More changes

+++ Assets/Scripts/Net/Data/ServerData.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 public class ServerData
 {
-    public Dictionary<uint, UserData> ConnectedUsers { get; private set; } = new Dictionary<uint, UserData>();
+    public Dictionary<ulong, UserData> ConnectedUsers { get; private set; } = new Dictionary<ulong, UserData>();
     public Dictionary<int, ServerLobby> ActiveLobbies { get; private set; } = new Dictionary<int, ServerLobby>();
 #if CNS_DEDICATED_SERVER_MULTI_LOBBY_AUTH
     public ServerSettings Settings { get; set; } = new ServerSettings();
